hdmitx: process the coverity [1/1]
PD#SWPL-219531

Problem:
process the coverity

Solution:
add protect

Verify:
s7

Test:
DRM-TX-79,DRM-TX-80,DRM-TX-81,DRM-TX-82

+++ drivers/media/vout/hdmi_tx_connector/hdmitx/tx20/hdmitx_scdc.c
@@ -42,7 +42,7 @@ static int scdc_ced_cnt(struct hdmitx_common *tx_comm)
 	struct ced_cnt *ced = &tx_comm->ced_cnt;
 	u8 raw[7];
 	u8 chksum;
-	int i;
+	u8 i;
 
 	memset(raw, 0, sizeof(raw));
 	memset(ced, 0, sizeof(struct ced_cnt));
